Add script and config to mirror regular APT repos
The script aptly-sync-dists.sh will read the configuration file matching
the name given as first parameter, then create the mirrors, snapshots
and publish a nightly version of the repositories.

This script is intended to mirror regular APT repositories with
multi-component support (those with the `dists` hierarchy of
directories).

The configuration files are provided as examples to mirror debian,
debian-security and ubuntu repositories.